@@ -1,5 +1,5 @@
 mod support;
-use support::*;
+use support::server;
 
 use std::io::Write;
 
@@ -91,7 +91,7 @@ async fn gzip_case(response_size: usize, chunk_size: usize) {
 
     let content: String = (0..response_size)
         .into_iter()
-        .map(|i| format!("test {}", i))
+        .map(|i| format!("test {i}"))
         .collect();
     let mut encoder = libflate::gzip::Encoder::new(Vec::new()).unwrap();
     match encoder.write(content.as_bytes()) {
