@@ -1,29 +1,30 @@
 #![cfg(not(target_arch = "wasm32"))]
 mod support;
 use futures_util::stream::StreamExt;
-use support::*;
+use hyper::Body;
+use support::server;
 
 #[tokio::test]
 async fn test_redirect_301_and_302_and_303_changes_post_to_get() {
     let client = reqwest::Client::new();
     let codes = [301u16, 302, 303];
 
-    for &code in codes.iter() {
+    for &code in &codes {
         let redirect = server::http(move |req| async move {
             if req.method() == "POST" {
-                assert_eq!(req.uri(), &*format!("/{}", code));
+                assert_eq!(req.uri(), &*format!("/{code}"));
                 http::Response::builder()
                     .status(code)
                     .header("location", "/dst")
                     .header("server", "test-redirect")
-                    .body(Default::default())
+                    .body(Body::default())
                     .unwrap()
             } else {
                 assert_eq!(req.method(), "GET");
 
                 http::Response::builder()
                     .header("server", "test-dst")
-                    .body(Default::default())
+                    .body(Body::default())
                     .unwrap()
             }
         });
@@ -44,22 +45,22 @@ async fn test_redirect_301_and_302_and_303_changes_post_to_get() {
 async fn test_redirect_307_and_308_tries_to_get_again() {
     let client = reqwest::Client::new();
     let codes = [307u16, 308];
-    for &code in codes.iter() {
+    for &code in &codes {
         let redirect = server::http(move |req| async move {
             assert_eq!(req.method(), "GET");
-            if req.uri() == &*format!("/{}", code) {
+            if req.uri() == &*format!("/{code}") {
                 http::Response::builder()
                     .status(code)
                     .header("location", "/dst")
                     .header("server", "test-redirect")
-                    .body(Default::default())
+                    .body(Body::default())
                     .unwrap()
             } else {
                 assert_eq!(req.uri(), "/dst");
 
                 http::Response::builder()
                     .header("server", "test-dst")
-                    .body(Default::default())
+                    .body(Body::default())
                     .unwrap()
             }
         });
@@ -81,7 +82,7 @@ async fn test_redirect_307_and_308_tries_to_post_again() {
     let _ = env_logger::try_init();
     let client = reqwest::Client::new();
     let codes = [307u16, 308];
-    for &code in codes.iter() {
+    for &code in &codes {
         let redirect = server::http(move |mut req| async move {
             assert_eq!(req.method(), "POST");
             assert_eq!(req.headers()["content-length"], "5");
@@ -89,19 +90,19 @@ async fn test_redirect_307_and_308_tries_to_post_again() {
             let data = req.body_mut().next().await.unwrap().unwrap();
             assert_eq!(&*data, b"Hello");
 
-            if req.uri() == &*format!("/{}", code) {
+            if req.uri() == &*format!("/{code}") {
                 http::Response::builder()
                     .status(code)
                     .header("location", "/dst")
                     .header("server", "test-redirect")
-                    .body(Default::default())
+                    .body(Body::default())
                     .unwrap()
             } else {
                 assert_eq!(req.uri(), "/dst");
 
                 http::Response::builder()
                     .header("server", "test-dst")
-                    .body(Default::default())
+                    .body(Body::default())
                     .unwrap()
             }
         });
@@ -123,10 +124,10 @@ async fn test_redirect_307_and_308_tries_to_post_again() {
 fn test_redirect_307_does_not_try_if_reader_cannot_reset() {
     let client = reqwest::blocking::Client::new();
     let codes = [307u16, 308];
-    for &code in codes.iter() {
+    for &code in &codes {
         let redirect = server::http(move |mut req| async move {
             assert_eq!(req.method(), "POST");
-            assert_eq!(req.uri(), &*format!("/{}", code));
+            assert_eq!(req.uri(), &*format!("/{code}"));
             assert_eq!(req.headers()["transfer-encoding"], "chunked");
 
             let data = req.body_mut().next().await.unwrap().unwrap();
@@ -136,7 +137,7 @@ fn test_redirect_307_does_not_try_if_reader_cannot_reset() {
                 .status(code)
                 .header("location", "/dst")
                 .header("server", "test-redirect")
-                .body(Default::default())
+                .body(Body::default())
                 .unwrap()
         });
 
@@ -166,7 +167,7 @@ async fn test_redirect_removes_sensitive_headers() {
             let mid_addr = rx.borrow().unwrap();
             assert_eq!(
                 req.headers()["referer"],
-                format!("http://{}/sensitive", mid_addr)
+                format!("http://{mid_addr}/sensitive")
             );
             http::Response::default()
         }
@@ -178,8 +179,8 @@ async fn test_redirect_removes_sensitive_headers() {
         assert_eq!(req.headers()["cookie"], "foo=bar");
         http::Response::builder()
             .status(302)
-            .header("location", format!("http://{}/end", end_addr))
-            .body(Default::default())
+            .header("location", format!("http://{end_addr}/end"))
+            .body(Body::default())
             .unwrap()
     });
 
@@ -205,7 +206,7 @@ async fn test_redirect_policy_can_return_errors() {
         http::Response::builder()
             .status(302)
             .header("location", "/loop")
-            .body(Default::default())
+            .body(Body::default())
             .unwrap()
     });
 
@@ -221,7 +222,7 @@ async fn test_redirect_policy_can_stop_redirects_without_an_error() {
         http::Response::builder()
             .status(302)
             .header("location", "/dont")
-            .body(Default::default())
+            .body(Body::default())
             .unwrap()
     });
 
@@ -247,7 +248,7 @@ async fn test_referer_is_not_set_if_disabled() {
             http::Response::builder()
                 .status(302)
                 .header("location", "/dst")
-                .body(Default::default())
+                .body(Body::default())
                 .unwrap()
         } else {
             assert_eq!(req.uri(), "/dst");
@@ -273,7 +274,7 @@ async fn test_invalid_location_stops_redirect_gh484() {
         http::Response::builder()
             .status(302)
             .header("location", "http://www.yikes{KABOOM}")
-            .body(Default::default())
+            .body(Body::default())
             .unwrap()
     });
 
@@ -283,6 +284,22 @@ async fn test_invalid_location_stops_redirect_gh484() {
 
     assert_eq!(res.url().as_str(), url);
     assert_eq!(res.status(), reqwest::StatusCode::FOUND);
+}
+
+#[tokio::test]
+async fn test_invalid_scheme_is_rejected() {
+    let server = server::http(move |_req| async move {
+        http::Response::builder()
+            .status(302)
+            .header("location", "htt://www.yikes.com/")
+            .body(Body::default())
+            .unwrap()
+    });
+
+    let url = format!("http://{}/yikes", server.addr());
+
+    let err = reqwest::get(&url).await.unwrap_err();
+    assert!(err.is_builder());
 }
 
 #[cfg(feature = "cookies")]
@@ -295,7 +312,7 @@ async fn test_redirect_302_with_set_cookies() {
                 .status(302)
                 .header("location", "/dst")
                 .header("set-cookie", "key=value")
-                .body(Default::default())
+                .body(Body::default())
                 .unwrap()
         } else {
             assert_eq!(req.uri(), "/dst");
@@ -325,7 +342,7 @@ async fn test_redirect_https_only_enforced_gh1312() {
         http::Response::builder()
             .status(302)
             .header("location", "http://insecure")
-            .body(Default::default())
+            .body(Body::default())
             .unwrap()
     });
 
