@@ -3,24 +3,35 @@
 // This is using the `tokio` runtime. You'll need the following dependency:
 //
 // `tokio = { version = "1", features = ["full"] }`
+#[cfg(feature = "http3")]
 #[cfg(not(target_arch = "wasm32"))]
 #[tokio::main]
 async fn main() -> Result<(), reqwest::Error> {
+    use http::Version;
+    use reqwest::{Client, IntoUrl, Response};
+
+    async fn get<T: IntoUrl + Clone>(url: T) -> reqwest::Result<Response> {
+        Client::builder()
+            .http3_prior_knowledge()
+            .build()?
+            .get(url)
+            .version(Version::HTTP_3)
+            .send()
+            .await
+    }
+
     // Some simple CLI args requirements...
-    let url = if let Some(url) = std::env::args().nth(1) {
-        url
-    } else {
-        println!("No CLI URL provided, using default.");
-        "https://hyper.rs".into()
+    let url = match std::env::args().nth(1) {
+        Some(url) => url,
+        None => {
+            println!("No CLI URL provided, using default.");
+            "https://hyper.rs".into()
+        }
     };
 
     eprintln!("Fetching {url:?}...");
 
-    // reqwest::get() is a convenience function.
-    //
-    // In most cases, you should create/build a reqwest::Client and reuse
-    // it for all requests.
-    let res = reqwest::get(url).await?;
+    let res = get(url).await?;
 
     eprintln!("Response: {:?} {}", res.version(), res.status());
     eprintln!("Headers: {:#?}\n", res.headers());
@@ -36,5 +47,5 @@ async fn main() -> Result<(), reqwest::Error> {
 // for wasm32 target, because tokio isn't compatible with wasm32.
 // If you aren't building for wasm32, you don't need that line.
 // The two lines below avoid the "'main' function not found" error when building for wasm32 target.
-#[cfg(target_arch = "wasm32")]
+#[cfg(any(target_arch = "wasm32", not(feature = "http3")))]
 fn main() {}
