@@ -1,7 +1,7 @@
 #![cfg(not(target_arch = "wasm32"))]
 mod support;
 use futures_util::stream::StreamExt;
-use support::*;
+use support::server;
 
 #[tokio::test]
 async fn text_part() {
